majooor updatees
cleaning is now listed as experimental and you will be prompted only once to activate or not, you can manually change the config to activate it later, setting the config is now its own javascript file and is passed arguments to set as the config settings.
last thing to do is clean up the cleaning process making it actually useable and able to remove whatever you desire from every line in the file not just the top.